requirements.txt
# Core ML
numpy>=1.24,<3
matplotlib>=3.7
scikit-learn>=1.3

# Embeddings
sentence-transformers>=2.2

# Document processing
pypdf>=4.0
python-docx>=1.0
PyMuPDF>=1.23
pillow>=10.0

# OCR (docTR with PyTorch backend)
python-doctr[torch]>=0.6

# Vector database
chromadb>=0.4

# LLM
anthropic>=0.18
openai>=1.0

# Web research
duckduckgo-search>=6.0

# Config
pyyaml>=6.0
python-dotenv>=1.0
pyotp>=2.9

# Web server
fastapi>=0.109
uvicorn[standard]>=0.27

# Google Calendar
google-auth>=2.27
google-auth-oauthlib>=1.2
google-api-python-client>=2.110

# Telegram bot
python-telegram-bot[job-queue]>=20.0
pytz

# Testing
pytest>=7.4
httpx>=0.26
